refactor(cache): share the timeseries wire types via @directus/types
cache.vue hand-mirrored the API's CacheTimeseriesBucket/CacheTimeseries as local
TimeseriesBucket/TimeseriesData interfaces, so an API-side field change (like the
six latency percentiles this stack just added to both) would silently drop in the
app. Move the three types to @directus/types (next to CacheFlushTarget, the same
pattern), re-export them from cache-events for internal API consumers, and import
them in the app — one owner, no drift.
